dashboard/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
markdown==3.5.1
